refactor: drop the execution step properties interface
This patch drops the ExecutionStepProperties interface. The properties
are now just a simple immutable final class. This gives us a cleaner
plan schema.
