+++ ksql-engine/src/test/java/io/confluent/ksql/structured/SchemaKTableTest.java
@@ -43,7 +43,7 @@ import io.confluent.ksql.execution.expression.tree.ColumnReferenceExp;
 import io.confluent.ksql.execution.expression.tree.ComparisonExpression;
 import io.confluent.ksql.execution.expression.tree.Expression;
 import io.confluent.ksql.execution.expression.tree.LongLiteral;
-import io.confluent.ksql.execution.plan.DefaultExecutionStepProperties;
+import io.confluent.ksql.execution.plan.ExecutionStepProperties;
 import io.confluent.ksql.execution.plan.ExecutionStep;
 import io.confluent.ksql.execution.plan.Formats;
 import io.confluent.ksql.execution.plan.JoinType;
@@ -204,7 +204,7 @@ public class SchemaKTableTest {
   private ExecutionStep buildSourceStep(final LogicalSchema schema, final KTable kTable) {
     final ExecutionStep sourceStep = Mockito.mock(ExecutionStep.class);
     when(sourceStep.getProperties()).thenReturn(
-        new DefaultExecutionStepProperties(schema, queryContext.getQueryContext()));
+        new ExecutionStepProperties(schema, queryContext.getQueryContext()));
     when(sourceStep.build(any())).thenReturn(
         KTableHolder.unmaterialized(kTable, schema, keySerdeFactory));
     return sourceStep;
